Fixed bugs on GPIO interrupt and sin and cos computation

+++ Core/Inc/Application/Working/working.h
@@ -3,9 +3,9 @@
 #define N_POINTS 		1024
 
 // Functions prototypes
-void computeSinCos(double *sin, double *cos, const int nPoints);
-void formatSin(double *signalToFormat, double *sin, double *cos);
-void convertToPWMlogic(uint32_t *freqArray, const double *fullSin, Signal sig, uint32_t size);
+void computeSinCos(float *sin, float *cos, const int nPoints);
+void formatSin(float *signalToFormat, float *sin, float *cos);
+void convertToPWMlogic(uint32_t *freqArray, const float *fullSin, Signal sig, uint32_t size);
